refactor : [refactor/#13/restful/1]
1. restful 규칙에 맞게 endpoint 변경
2. OrderController 와 StoreController 분리
3. OrderService 와 StoreService 분리
4. http 파일 삭제
